feat: add controller-input extensions
Added a new class with extension methods for simulating input from
analog axes on game controllers. One pair of methods is used to
simulate two-sided axes (like the x axis of a thumbstick), and is
compatible with Godot's Input.GetAxis() method. The other pair of
methods is used to simulate one-sided axes (like a gamepad trigger),
and is compatible with Godot's Input.GetActionStrength() method.